dashboard_Dash/requirements.txt
attrs==19.3.0
Brotli==1.0.7
certifi==2020.6.20
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
dash==1.13.4
dash-bootstrap-components==0.10.3
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
Fiona==1.8.13.post1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
geopandas==0.8.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
munch==2.5.0
mysql-connector==2.2.9
numpy==1.19.0
pandas==1.0.5
plotly==4.8.1
pyproj==2.6.1.post1
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
Shapely==1.7.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
